Do `symbol->function` conversion in `call` instead of `call-help`
`call-help` gets called recursively a few times, and the F argument
is not always a symbol. Handling this in `call` is cleaner and avoids
unnecessary calls to `symbol->function`.

The `function-abstraction` helper used in the `function` macro was
modified to use `lambda` instead of `/.`. The purpose of this is
to be able to use the more lighweight `eval-kl` function in the
resulting code instead of `eval`. Being that `symbol->function`
whould be fast, it is better to avoid `eval`.

`symbol->function` now takes an arity parameter so that it can be
used to wrap symbols referring to functions not defined yet.